requirements.txt
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.28.5
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.5
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
tqdm==4.62.3
